Refactor Phantom rotation and update data types
- Renamed `PhantomDefault` to `Phantom Jobs Loaded`.
- Added `AttackAbility` method for player status checks.
- Enhanced `HealSingleAbility` with `PhantomJudgmentPvE` check.
- Changed nearby hostiles counting in `SCH_Default` from float to integer.
- Updated `AnimationLockTime` and `TimeToKill` properties to integers across multiple classes.
- Refactored `GetTTK` method to return integers.
- Modified `BaseAction` to use static `Player.AnimationLock`.
- Adjusted various JSON files for data structure changes.
- Updated `RotationConfigWindow` to handle new integer types.
- Added `DisplayStatus` method in `DutyRotation` for ability status.
- Enhanced modification methods to include new status checks.
- Refactored `IsPlayerOccupied` for better readability.
- Improved ratio calculations in `StateUpdater` by switching to integer comparisons.

+++ RotationSolver/Updaters/StateUpdater.cs
@@ -269,7 +269,7 @@ internal static class StateUpdater
 
         if (DataCenter.Role == JobRole.Tank)
         {
-            bool movingHere = (float)DataCenter.NumberOfHostilesInRange / DataCenter.NumberOfHostilesInMaxRange > 0.3f;
+            bool movingHere = DataCenter.NumberOfHostilesInRange / DataCenter.NumberOfHostilesInMaxRange > 0.3f;
 
             int tarOnMeCount = 0;
             int attackedCount = 0;
@@ -285,7 +285,7 @@ internal static class StateUpdater
                 }
             }
 
-            bool attacked = (float)attackedCount / tarOnMeCount > 0.7f;
+            bool attacked = attackedCount / tarOnMeCount > 0.7f;
 
             if (tarOnMeCount >= Service.Config.AutoDefenseNumber
                 && Player.Object.GetHealthRatio() <= Service.Config.HealthForAutoDefense
